requirements.txt
openai
aiohttp
numpy
fastapi
uvicorn
python-dotenv
pydantic
opencv-python
python-multipart
requests
python-multipart
